+++ input_manager.py
@@ -40,8 +40,9 @@ class InputManager:
 
         # 上下文按键别名映射
         self.CONTEXT_ALIAS = {
-            "START": "MENU", 
-            "GAME": "GAME", 
+            "START": "MENU",
+            "LOBBY": "MENU",
+            "GAME": "GAME",
             "PAUSE": "MENU"
             }